server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
pymongo==4.5.0
bcrypt==4.0.1
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0 
requests==2.31.0
urllib3==1.26.16
cryptography==41.0.4